Move common action_class elements to _base partial
