+++ resources/config_auto_auth.rb
@@ -36,24 +36,19 @@ property :vault_mode, [String, Symbol],
           desired_state: false,
           description: 'Vault service operation mode. Defaults to agent.'
 
-action_class do
-  include Vault::Cookbook::Helpers
-  include Vault::Cookbook::ResourceHelpers
-end
-
 load_current_value do
   case entry_type
   when :method
     option_data = vault_hcl_config_current_load(config_file).dig(vault_hcl_config_type, entry_type.to_s, type)
 
-    current_value_does_not_exist! if option_data.nil?
+    current_value_does_not_exist! if nil_or_empty?(option_data)
 
     options option_data
   when :sink
     option_data = vault_hcl_config_current_load(config_file).dig(vault_hcl_config_type, entry_type.to_s) || []
     option_data = option_data.filter { |s| s.dig(type, 'config', 'path').eql?(path) }
 
-    current_value_does_not_exist! if option_data.empty?
+    current_value_does_not_exist! if nil_or_empty?(option_data)
     raise Chef::Exceptions::InvalidResourceReference,
           "Filter matched #{option_data.count} auto_auth #{entry_type} configuration items but only should match one." if option_data.count > 1
